refactor : applying new clear arch structure
